+++ infra/resources/_modules/key_vault/variables.tf
@@ -1,0 +1,70 @@
+######################
+#  Common Variables  #
+######################
+
+variable "prefix" {
+  type    = string
+  default = "io"
+  validation {
+    condition = (
+      length(var.prefix) < 6
+    )
+    error_message = "Max length is 6 chars."
+  }
+}
+
+variable "env_short" {
+  type = string
+  validation {
+    condition = (
+      length(var.env_short) <= 1
+    )
+    error_message = "Max length is 1 chars."
+  }
+}
+
+variable "location" {
+  type        = string
+  description = "Azure region"
+}
+
+variable "location_short" {
+  type        = string
+  description = "Azure region"
+}
+
+variable "domain" {
+  type        = string
+  description = "Domain name of the application"
+}
+
+variable "tags" {
+  type        = map(any)
+  description = "Resource tags"
+}
+
+variable "resource_group_name" {
+  type        = string
+  description = "Resource group name for the Function App services"
+}
+
+variable "tenant_id" {
+  type        = string
+  description = "Id of the Tenant to deploy the KeyVault"
+}
+
+
+variable "peps_snet_id" {
+  type        = string
+  description = "Id of the subnet which holds private endpoints"
+}
+
+variable "vault_private_dns_zone_id" {
+  type        = string
+  description = "Id of the KeyVault Private DNS Zone"
+}
+
+variable "apim_principal_id" {
+  type        = string
+  description = "Id of the APIM-assigned managed identity"
+}