+++ infra/resources/prod/providers.tf
@@ -1,0 +1,24 @@
+terraform {
+  backend "azurerm" {
+    resource_group_name  = "terraform-state-rg"
+    storage_account_name = "iopitntfst001"
+    container_name       = "terraform-state"
+    key                  = "io-ipatente.resources.prod.tfstate"
+  }
+
+  required_providers {
+    azurerm = {
+      source  = "hashicorp/azurerm"
+      version = "<= 3.116.0"
+    }
+    azuread = {
+      source  = "hashicorp/azuread"
+      version = "2.48.0"
+    }
+  }
+}
+
+provider "azurerm" {
+  features {}
+  storage_use_azuread = true
+}