+++ packages/core/lib/index.ts
@@ -1,1 +1,4 @@
+export * from "./auth";
 export * from "./bff";
+export * from "./config";
+export * from "./middleware";