+++ packages/core/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @io-ipatente/core
 
+## 0.4.1
+
+### Patch Changes
+
+- 937ee93: Set Trust Host AuthJS conf
+
+## 0.4.0
+
+### Minor Changes
+
+- f67d2b9: fix authentication flow in dev mode
+
+## 0.3.0
+
+### Minor Changes
+
+- 6fd5a08: Move middleware and auth config to @io-ipatente/core
+
+## 0.2.0
+
+### Minor Changes
+
+- 738cc18: move config to @io-ipatente/core
+- fa97d5e: Added interop msw mock
+
 ## 0.1.0
 
 ### Minor Changes